requirements.txt

certifi==2016.2.28
cloudpickle==0.8.0
cycler==0.10.0
dask==1.1.4
decorator==4.3.2
kiwisolver==1.0.1
matplotlib==3.0.3
networkx==2.2
numpy==1.16.2
opencv-python==4.0.0.21
openslide-python==1.1.1
pandas==0.24.1
Pillow==5.4.1
protobuf==3.7.0
pyparsing==2.3.1
python-dateutil==2.8.0
pytz==2018.9
PyWavelets==1.0.2
scikit-image==0.14.2
scipy==1.2.1
six==1.12.0
tensorboardX==1.6
toolz==0.9.0
torch==1.0.1.post2
torchvision==0.2.2.post3
